The bot was abandoned
"en: As of today (2017-07-19), I've decided to abandon this project. Creating a bot for pixelcanvas.io has become increasingly difficult, although still possible. The code will stay here in case it helps anyone. But I lack the motivation and time to maintain this. If you are looking for an actively developed bot, you should take a look at RogerioBlanco/PixelCanvasBot, it looks actively developed."